se elimino metodo find de atraccion usado por deleteLogico y se optimizo dicho metodo

+++ src/main/java/controller/attractions/DeleteAttractionServlet.java
@@ -38,7 +38,7 @@ public class DeleteAttractionServlet extends HttpServlet {
 	 */
 	@Override
 	protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-		Long id = Long.parseLong(req.getParameter("id"));
+		Integer id = Integer.parseInt(req.getParameter("id"));
 		
 		try {
 			attractionService.delete(id);
